move jmx mapping support from agent to jvm ext
The ability to query JMX based on a configured mapping is
useful in regular apps that may be using the spectator
directly rather than using the agent. This change allows
the mappings to be easily used in either context.
