+++ spectator-ext-jvm/build.gradle
@@ -1,5 +1,6 @@
 dependencies {
   compileApi project(':spectator-api')
+  compile 'com.typesafe:config'
 }
 
 jar {
