+++ spectator-ext-jvm/src/main/java/com/netflix/spectator/jvm/JmxMeter.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014-2017 Netflix, Inc.
+ * Copyright 2014-2018 Netflix, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.netflix.spectator.agent;
+package com.netflix.spectator.jvm;
 
 import com.netflix.spectator.api.Id;
 import com.netflix.spectator.api.Measurement;
