+++ spectator-ext-jvm/src/main/resources/artifactory.conf
