+++ spectator-ext-jvm/src/main/resources/bitbucket.conf
