+++ spectator-ext-jvm/src/main/resources/cassandra.conf
