+++ spectator-ext-jvm/src/main/resources/tomcat.conf
